# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vehicle Management System Api Documentation",
        "description": "L5 Swagger API Documentation for Laravel Project Pharmacy",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/trucks": {
            "get": {
                "tags": [
                    "Trucks"
                ],
                "summary": "Get List of all Trucks",
                "description": "Returns List of all trucks",
                "operationId": "truckIndex",
                "responses": {
                    "200": {
                        "description": "A list of trucks"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page Not Found"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Trucks"
                ],
                "summary": "Store a truck",
                "description": "Store truck",
                "operationId": "truckStore",
                "requestBody": {
                    "description": "Pass truck data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "make",
                                    "model",
                                    "year",
                                    "seats",
                                    "color",
                                    "vin",
                                    "current_mileage",
                                    "service_interval",
                                    "next_service"
                                ],
                                "properties": {
                                    "make": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "KIA"
                                    },
                                    "model": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Sportage"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "2023"
                                    },
                                    "seats": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "4"
                                    },
                                    "color": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "White"
                                    },
                                    "vin": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "KUCHBELIKH465452"
                                    },
                                    "current_mileage": {
                                        "type": "decimal",
                                        "format": "text",
                                        "example": "11.5"
                                    },
                                    "service_interval": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "3000KM"
                                    },
                                    "next_service": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "6000KM"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Display a truck"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page Not Found "
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/trucks/{id}": {
            "get": {
                "tags": [
                    "Trucks"
                ],
                "summary": "Show a Truck",
                "description": "Display a truck",
                "operationId": "truckShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the truck to show",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "show a truck"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page Not Found"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Trucks"
                ],
                "summary": "Update a truck",
                "description": "Update truck",
                "operationId": "truckUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the truck to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass truck data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "make",
                                    "model",
                                    "year",
                                    "seats",
                                    "color",
                                    "vin",
                                    "current_mileage",
                                    "service_interval",
                                    "next_service"
                                ],
                                "properties": {
                                    "make": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "KIA"
                                    },
                                    "model": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Sportage"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "2023"
                                    },
                                    "seats": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "4"
                                    },
                                    "color": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "White"
                                    },
                                    "vin": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "KUCHBELIKH465452"
                                    },
                                    "current_mileage": {
                                        "type": "decimal",
                                        "format": "text",
                                        "example": "11.5"
                                    },
                                    "service_interval": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "3000KM"
                                    },
                                    "next_service": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "6000KM"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a truck"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page Not Found "
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Trucks"
                ],
                "summary": "Get a Truck",
                "description": "Delete a truck",
                "operationId": "truckDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the truck to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete a truck"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page Not Found"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}